[FLINK-12777][network] Extract CheckpointBarrierAligner from BarrierBuffer
